requirements.txt
albumentations==1.3.0
bbox-visualizer==0.1.0
blobfile==2.0.2
cython==3.0.0
einops==0.6.1
ensemble-boxes==1.0.9
ftfy==6.1.1
imageio==2.28.1
joblib==1.2.0
numpy==1.23.5
matplotlib==3.7.1
mypy==1.5.1
opencv-python-headless==4.7.0.72
pandas==2.0.1
pathtools==0.1.2
Pillow==9.1.0
protobuf==4.23.0
pycocotools==2.0.6
PyWavelets==1.4.1
PyYAML==6.0
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
seaborn==0.12.2
shapely==1.8.5
tensorboard==2.13.0
timm==0.5.4
triton==2.0.0
tqdm==4.65.0
wandb==0.15.2
